backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-tailor-api"
version = "0.1.0"
description = "CV & Cover-Letter Auto-Tailor Backend API"
authors = [
    {name = "CV Tailor Team", email = "team@cv-tailor.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.68.0",
    "celery>=5.5.3",
    "django>=4.2.24",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "jinja2>=3.1.6",
    "openai>=1.108.2",
    "pypdf2>=3.0.1",
    "python-decouple>=3.8",
    "python-docx>=1.2.0",
    "qrcode[pil]>=8.2",
    "redis>=6.4.0",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cv_tailor"]
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14,<4.0",
    "celery>=5.3,<6.0",
    "redis>=5.0,<6.0",
    "psycopg2-binary>=2.9,<3.0",
    "gunicorn>=21.0,<22.0",
    "whitenoise>=6.5,<7.0",
    "django-cors-headers>=4.3,<5.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-ratelimit>=4.1,<5.0",
    "django-extensions>=3.2,<4.0",
    "pillow>=10.0,<11.0",
    "requests>=2.31,<3.0",
    "python-decouple>=3.8,<4.0",
    "openai>=1.0,<2.0",
    "anthropic>=0.7,<1.0",
    "reportlab>=4.0,<5.0",
    "python-docx>=1.1,<2.0",
    "pypdf2>=3.0,<4.0",
    "jinja2>=3.1,<4.0",
    "pydantic>=2.0,<3.0",
    "httpx>=0.25,<1.0",
    "aiohttp>=3.8,<4.0",
    "numpy>=1.24,<2.0",
    "pandas>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-django>=4.5,<5.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-asyncio>=0.21,<1.0",
    "black>=23.0,<24.0",
    "isort>=5.12,<6.0",
    "flake8>=6.0,<7.0",
    "mypy>=1.5,<2.0",
    "django-debug-toolbar>=4.2,<5.0",
    "factory-boy>=3.3,<4.0",
    "pre-commit>=3.4,<4.0",
    "ipython>=8.0,<9.0",
]
prod = [
    "sentry-sdk>=1.30,<2.0",
    "django-health-check>=3.17,<4.0",
    "django-prometheus>=2.3,<3.0",
]
ml = [
    "scikit-learn>=1.3,<2.0",
    "transformers>=4.30,<5.0",
    "sentence-transformers>=2.2,<3.0",
    "mlflow>=2.6,<3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cv_tailor.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "manage.py",
    "*/settings/*",
    "*/__pycache__/*",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "black>=23.0",
    "mypy>=1.5",
    "django-debug-toolbar>=4.2",
]
